If timestamp is not available for none of the containers, skip it:
If timestamp is not available for any containers, it means that the
docker remote api doesn't give any data. So it will be a row of NaN's.
If we include NaN for timestamp, it will cause problems when reporting data
while reading the timestamp since it expects long and not NaN string.